pingce/requirements.txt
Flask==2.0.1
openai==1.3.5
python-dotenv==0.21.0